COMMERCE-1453 Move SimpleCPTypeConstants to api module. Remove Suppression
