+++ source-formatter-suppressions.xml
@@ -10,7 +10,6 @@
 		<suppress checks="BNDExportsCheck" files="commerce-initializer-util/bnd\.bnd" />
 		<suppress checks="BNDExportsCheck" files="commerce-product-definitions-web/bnd\.bnd" />
 		<suppress checks="BNDExportsCheck" files="commerce-product-service/bnd\.bnd" />
-		<suppress checks="BNDExportsCheck" files="commerce-product-type-simple/bnd\.bnd" />
 		<suppress checks="BNDExportsCheck" files="commerce-service/bnd\.bnd" />
 		<suppress checks="BNDExportsCheck" files="commerce-shipping-web/bnd\.bnd" />
 		<suppress checks="GradleTestDependencyVersionCheck" files=".*" />
